+++ application/cloudevent-converter/jackson/src/test/java/org/occurrent/application/converter/jackson/JacksonCloudEventConverterTest.java
@@ -15,9 +15,10 @@
  *  limitations under the License.
  */
 
-package org.occurrent.application.converter.xstream;
+package org.occurrent.application.converter.jackson;
 
-import com.thoughtworks.xstream.XStream;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
 import org.junit.jupiter.api.Test;
@@ -31,14 +32,13 @@ import java.time.temporal.ChronoUnit;
 import java.util.Date;
 import java.util.UUID;
 
-import static java.nio.charset.StandardCharsets.UTF_16LE;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.time.ZoneOffset.UTC;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.within;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-public class XStreamCloudEventConverterTest {
+public class JacksonCloudEventConverterTest {
 
     private static final URI CLOUD_EVENT_SOURCE = URI.create("urn:occurrent:domain");
 
@@ -46,8 +46,8 @@ public class XStreamCloudEventConverterTest {
     @Test
     void converts_domain_event_to_cloud_event_using_default_configuration() {
         // Given
-        XStream xStream = new XStream();
-        XStreamCloudEventConverter<DomainEvent> cloudEventConverter = new XStreamCloudEventConverter<>(xStream, CLOUD_EVENT_SOURCE);
+        ObjectMapper objectMapper = new ObjectMapper();
+        JacksonCloudEventConverter<DomainEvent> cloudEventConverter = new JacksonCloudEventConverter<>(objectMapper, CLOUD_EVENT_SOURCE);
         NameDefined domainEvent = new NameDefined(UUID.randomUUID().toString(), new Date(), "name");
 
         // When
@@ -56,13 +56,13 @@ public class XStreamCloudEventConverterTest {
         // Then
         assertAll(
                 () -> assertThat(cloudEvent.getId()).matches("([a-f0-9]{8}(-[a-f0-9]{4}){4}[a-f0-9]{8})"),
-                () -> assertThat(cloudEvent.getType()).isEqualTo(NameDefined.class.getSimpleName()),
+                () -> assertThat(cloudEvent.getType()).isEqualTo(NameDefined.class.getName()),
                 () -> assertThat(cloudEvent.getSource()).isEqualTo(CLOUD_EVENT_SOURCE),
                 () -> assertThat(cloudEvent.getSubject()).isNull(),
                 () -> assertThat(cloudEvent.getTime()).isCloseToUtcNow(within(3, ChronoUnit.SECONDS)),
                 () -> assertThat(cloudEvent.getDataSchema()).isNull(),
-                () -> assertThat(cloudEvent.getDataContentType()).isEqualTo("application/xml"),
-                () -> assertThat(new String(cloudEvent.getData().toBytes(), UTF_8)).isEqualTo(xStream.toXML(domainEvent))
+                () -> assertThat(cloudEvent.getDataContentType()).isEqualTo("application/json"),
+                () -> assertThat(new String(cloudEvent.getData().toBytes(), UTF_8)).isEqualTo(objectMapper.writeValueAsString(domainEvent))
         );
     }
 
@@ -70,9 +70,8 @@ public class XStreamCloudEventConverterTest {
     @Test
     void converts_domain_event_to_cloud_event_using_non_default_configuration() {
         // Given
-        XStream xStream = new XStream();
-        XStreamCloudEventConverter<DomainEvent> cloudEventConverter = new XStreamCloudEventConverter.Builder<DomainEvent>(xStream, CLOUD_EVENT_SOURCE)
-                .charset(UTF_16LE)
+        ObjectMapper objectMapper = new ObjectMapper();
+        JacksonCloudEventConverter<DomainEvent> cloudEventConverter = new JacksonCloudEventConverter.Builder<DomainEvent>(objectMapper, CLOUD_EVENT_SOURCE)
                 .contentType("application/xstream+xml")
                 .idMapper(DomainEvent::getEventId)
                 .subjectMapper(__ -> "subject")
@@ -94,29 +93,24 @@ public class XStreamCloudEventConverterTest {
                 () -> assertThat(cloudEvent.getTime()).isEqualTo(OffsetDateTime.ofInstant(domainEvent.getTimestamp().toInstant(), UTC)),
                 () -> assertThat(cloudEvent.getDataSchema()).isNull(),
                 () -> assertThat(cloudEvent.getDataContentType()).isEqualTo("application/xstream+xml"),
-                () -> assertThat(new String(cloudEvent.getData().toBytes(), UTF_16LE)).isEqualTo(xStream.toXML(domainEvent))
+                () -> assertThat(new String(cloudEvent.getData().toBytes(), UTF_8)).isEqualTo(objectMapper.writeValueAsString(domainEvent))
         );
     }
 
     @Test
     void converts_cloud_event_to_domain_event() {
         // Given
-        XStream xStream = new XStream();
-        xStream.allowTypeHierarchy(DomainEvent.class);
-        XStreamCloudEventConverter<DomainEvent> cloudEventConverter = new XStreamCloudEventConverter<>(xStream, CLOUD_EVENT_SOURCE);
+        ObjectMapper objectMapper = new ObjectMapper();
+        JacksonCloudEventConverter<DomainEvent> cloudEventConverter = new JacksonCloudEventConverter<>(objectMapper, CLOUD_EVENT_SOURCE);
 
-        String xml = "<org.occurrent.domain.NameDefined>\n" +
-                "  <eventId>45a5925b-b1df-41f2-b51b-d5bf6d0fba88</eventId>\n" +
-                "  <timestamp>2021-09-24 10:39:03.556 UTC</timestamp>\n" +
-                "  <name>name</name>\n" +
-                "</org.occurrent.domain.NameDefined>";
+        String json = "{\"eventId\":\"6ec13db8-b216-41cc-9e61-d4553cfc5476\",\"timestamp\":1632482491299,\"name\":\"name\"}";
 
         CloudEvent cloudEvent = CloudEventBuilder.v1()
                 .withId(UUID.randomUUID().toString())
                 .withSource(CLOUD_EVENT_SOURCE)
-                .withType(NameDefined.class.getSimpleName())
+                .withType(NameDefined.class.getName())
                 .withTime(OffsetDateTime.now())
-                .withData(xml.getBytes(UTF_8))
+                .withData(json.getBytes(UTF_8))
                 .build();
 
         // When
@@ -127,8 +121,8 @@ public class XStreamCloudEventConverterTest {
         NameDefined nameDefined = (NameDefined) domainEvent;
         assertAll(
                 () -> assertThat(nameDefined.getName()).isEqualTo("name"),
-                () -> assertThat(nameDefined.getEventId()).isEqualTo("45a5925b-b1df-41f2-b51b-d5bf6d0fba88"),
-                () -> assertThat(nameDefined.getTimestamp()).isEqualTo(new Date(LocalDateTime.of(2021, 9, 24, 10, 39, 3, 556_000_000).atZone(UTC).toInstant().toEpochMilli()))
+                () -> assertThat(nameDefined.getEventId()).isEqualTo("6ec13db8-b216-41cc-9e61-d4553cfc5476"),
+                () -> assertThat(nameDefined.getTimestamp()).isEqualTo(new Date(LocalDateTime.of(2021, 9, 24, 11, 21, 31, 299_000_000).atZone(UTC).toInstant().toEpochMilli()))
         );
     }
 }